+++ lib/src/network/port_number_validator.dart
@@ -1,9 +1,10 @@
 import '../../localization/l10n.dart';
 import '../base_validator.dart';
 
-class IntegerValidator extends BaseValidator<String> {
-  IntegerValidator({
-    int? this.radix,
+class PortNumberValidator extends BaseValidator<String> {
+  PortNumberValidator({
+    this.min = 0,
+    this.max = 65535,
 
     /// {@macro base_validator_error_text}
     super.errorText,
@@ -12,16 +13,18 @@ class IntegerValidator extends BaseValidator<String> {
     super.checkNullOrEmpty,
   });
 
-  int? radix;
+  final int min;
+
+  final int max;
 
   @override
   String get translatedErrorText =>
-      FormBuilderLocalizations.current.integerErrorText;
+      FormBuilderLocalizations.current.portNumberErrorText(min, max);
 
   @override
   String? validateValue(String? valueCandidate) {
-    final int? number = int.tryParse(valueCandidate!, radix: radix);
-    if (number == null) {
+    final int? port = int.tryParse(valueCandidate!);
+    if (port == null || port < min || port > max) {
       return errorText;
     }
     return null;